Backend/requirements.txt
celery==5.4.0
Django==5.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
redis==5.0.8
requests==2.32.3
gunicorn==23.0.0